TEMP: (will remove commit): Refactored leaders list block, exposed leaders in graphql
