+++ frontend/src/types/leaders.ts
@@ -1,3 +1,7 @@
 export type LeadersListProps = {
   leaders: string
 }
+
+export type LeadersListBlockProps = {
+  [key: string]: string
+}
